Update conditional for getting the value
Hello, was getting a notice (with debug turned on) if say a post using this as a ACF metabox didn't have an address or coordinates value (and wasn't a required field). Updating this conditional fixed that and now properly displays nothing if empty or null. Props to https://github.com/codearachnid for helping me solve this.